Make <select> special in innerText
`select` is a replaced element, but for Web compat we need
to include the text of `option` elements in innerText.
The new text is intended to match the behavior of Gecko,
which in turn is very close to the legacy IE/Edge behavior.

Fixes https://github.com/rocallahan/innerText-spec/issues/5